Split up the need-reducer.
